requirements.txt
pywin32
pandas
numpy
python-dotenv
schedule
PyQt5 